Integrate dependency injection and seperation of the command line and libraries
