+++ ReplaceInFiles/IFileSearcher.cs
@@ -1,0 +1,13 @@
+﻿namespace ReplaceInFiles
+{
+    public interface IFileSearcher
+    {
+        bool HasExtensionDefined();
+        FileSearcher IgnoreFolderNames(params string[] folderNames);
+        FileSearcher IncludeSubfolders(bool searchSubfolders);
+        FileSearcher InDirectory(string directory);
+        FileSearcher ParallelsExecution(int parallelsExecution);
+        List<string> Search();
+        FileSearcher WithExtensions(params string[] extensions);
+    }
+}
