+++ ReplaceInFiles/IFolderSearcher.cs
