+++ ReplaceInFilesTests/ReplaceInFilesTests.csproj
@@ -17,7 +17,7 @@
     <PackageReference Include="Serilog" Version="3.0.1" />
     <PackageReference Include="Serilog.Sinks.Console" Version="4.1.0" />
     <PackageReference Include="System.IO.Abstractions.TestingHelpers" Version="19.2.69" />
-    <PackageReference Include="xunit" Version="2.6.0" />
+    <PackageReference Include="xunit" Version="2.6.1" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.5.3">
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
